refactor: mission detail components
