+++ src/features/mission/pages/index.ts
@@ -1,2 +1,2 @@
-export { MissionsPage } from './MissionsPage';
-export { MissionDetailPage } from './MissionDetailPage';
+export { MissionsPage } from "./MissionsPage";
+export { MissionDetailPage } from "./MissionDetailPage";
